requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
easyocr==1.7.2
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.13.1
fsspec==2024.6.1
gradio==5.37.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
joblib==1.5.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
opencv-python-headless==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.3
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==70.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.13.3
tifffile==2025.6.11
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
celery==5.4.0 
flower==2.0.1 
redis==5.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
